requirements.txt
crewai>=0.11.0
openai>=1.13.3
jira>=3.5.1
python-dotenv>=1.0.0
langchain-openai>=0.0.2